Report installed PEcAn version

Reports the currently installed or loaded version(s) of each PEcAn package, as well as the versions expected by the specified release(s) of the PEcAn system.

Usage

pecan_version(version = max(PEcAn.all::pecan_releases$version), exact = FALSE)

Arguments

version

PEcAn release number to use for expected package versions

exact

Show only tags that exactly match version, or all tags that have it as a substring?

Value

data frame with columns for package name, expected version(s), installed version, and Git hash (if known). If the sessioninfo package is installed, an additional column reports where each package was installed from: local, github, CRAN, etc.

Details

When exact = FALSE (the default), pecan_version will show all releases whose names contain version as a substring. This is mostly so that e.g. "1.4.2" finds the release tag "v1.4.2", but since all matches are returned it can also be handy for seeing all patches sharing a minor version number (see examples).

If more than one version of a package is installed, the output from pecan_version() will contain multiple rows for that package. This can occur if you have installed different versions to different locations in .libPaths(), or if you've loaded a new version into your current session by loading it from its source directory without installing it to the R library. If you see multiple rows unexpectedly, try find.package(<pkgname>, verbose = TRUE) to see where each version was found.

Examples

pecan_version()
#>  package               v1.10.0 installed   source              
#>  PEcAn.BASGRA          1.8.2   1.8.2       local (/pecan/mod...
#>  PEcAn.BIOCRO          1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.CABLE           1.7.5   NA          NA                  
#>  PEcAn.CLM45           1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.DALEC           1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.DB              1.8.2   1.8.2.9000  local (/pecan/bas...
#>  PEcAn.ED2             1.8.2   1.8.2       local (/pecan/mod...
#>  PEcAn.FATES           1.8.1   1.8.1       local (/pecan/mod...
#>  PEcAn.GDAY            1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.JULES           1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.LDNDC           1.0.2   1.0.2       local (/pecan/mod...
#>  PEcAn.LINKAGES        1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.LPJGUESS        1.9.0   1.9.0       local (/pecan/mod...
#>  PEcAn.MA              1.7.5   1.7.5.9000  local (/pecan/mod...
#>  PEcAn.MAAT            1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.MAESPA          1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.ModelName       0.0.1   0.0.1       local (/pecan/mod...
#>  PEcAn.PRELES          1.7.5   NA          NA                  
#>  PEcAn.RothC           <NA>    0.0.0.9000  local (/pecan/mod...
#>  PEcAn.SIBCASA         0.0.3   0.0.3       local (/pecan/mod...
#>  PEcAn.SIPNET          1.10.0  1.10.0      local (/pecan/mod...
#>  PEcAn.STICS           1.8.2   1.8.2       local (/pecan/mod...
#>  PEcAn.all             1.10.0  1.10.0      local               
#>  PEcAn.allometry       1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.assim.batch     1.9.1   1.9.1       local (/pecan/mod...
#>  PEcAn.benchmark       1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.data.atmosphere 1.9.1   1.9.1.9000  local (/pecan/mod...
#>  PEcAn.data.land       1.9.0   1.9.0       local (/pecan/mod...
#>  PEcAn.data.mining     1.7.5   NA          NA                  
#>  PEcAn.data.remote     1.9.1   1.9.1       local (/pecan/mod...
#>  PEcAn.dvmdostem       1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.emulator        1.8.2   1.8.2       local (/pecan/mod...
#>  PEcAn.logger          1.8.4   1.8.4       local (/pecan/bas...
#>  PEcAn.photosynthesis  1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.priors          1.7.5   1.7.5       local (/pecan/mod...
#>  PEcAn.qaqc            1.7.5   1.7.5       local (/pecan/bas...
#>  PEcAn.remote          1.9.1   1.9.1       local (/pecan/bas...
#>  PEcAn.settings        1.9.1   1.9.1.9000  local (/pecan/bas...
#>  PEcAn.uncertainty     1.9.0   1.9.0.9000  local (/pecan/mod...
#>  PEcAn.utils           1.8.2   1.8.2       local (/pecan/bas...
#>  PEcAn.visualization   1.8.2   1.8.2       local (/pecan/bas...
#>  PEcAn.workflow        1.10.0  1.10.0      local (/pecan/bas...
#>  PEcAnAssimSequential  1.10.0  1.10.0.9000 local (/pecan/mod...
#>  PEcAnRTM              1.9.1   1.9.1       local (/pecan/mod...
pecan_version(c("1.4.0", "1.5.0", "1.6.0"))
#>  package                v1.4.0 v1.5.0 v1.6.0 installed   source              
#>  PEcAn.BASGRA           <NA>   <NA>   <NA>   1.8.2       local (/pecan/mod...
#>  PEcAn.BIOCRO           1.4.0  1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.CABLE            <NA>   <NA>   1.6.0  NA          NA                  
#>  PEcAn.CLM45            <NA>   1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.DALEC            1.4.0  1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.DB               1.4.0  1.5.0  1.6.0  1.8.2.9000  local (/pecan/bas...
#>  PEcAn.ED2              1.4.0  1.5.0  1.6.0  1.8.2       local (/pecan/mod...
#>  PEcAn.FATES            <NA>   1.5.0  1.6.0  1.8.1       local (/pecan/mod...
#>  PEcAn.GDAY             <NA>   1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.JULES            <NA>   1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.LDNDC            <NA>   <NA>   <NA>   1.0.2       local (/pecan/mod...
#>  PEcAn.LINKAGES         <NA>   1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.LPJGUESS         <NA>   1.5.0  1.6.0  1.9.0       local (/pecan/mod...
#>  PEcAn.MA               1.4.0  1.5.0  1.6.0  1.7.5.9000  local (/pecan/mod...
#>  PEcAn.MAAT             <NA>   1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.MAESPA           <NA>   1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.ModelName        1.4.0  1.5.0  1.6.0  0.0.1       local (/pecan/mod...
#>  PEcAn.PRELES           <NA>   1.5.0  1.6.0  NA          NA                  
#>  PEcAn.RothC            <NA>   <NA>   <NA>   0.0.0.9000  local (/pecan/mod...
#>  PEcAn.SIBCASA          <NA>   <NA>   <NA>   0.0.3       local (/pecan/mod...
#>  PEcAn.SIPNET           1.4.0  1.5.0  1.6.0  1.10.0      local (/pecan/mod...
#>  PEcAn.STICS            <NA>   <NA>   <NA>   1.8.2       local (/pecan/mod...
#>  PEcAn.all              1.4.0  1.5.0  1.6.0  1.10.0      local               
#>  PEcAn.allometry        1.4.0  1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.assim.batch      1.4.0  1.5.0  1.6.0  1.9.1       local (/pecan/mod...
#>  PEcAn.assim.sequential 1.4.0  1.5.0  1.6.0  NA          NA                  
#>  PEcAn.benchmark        1.4.0  1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.data.atmosphere  1.4.0  1.5.0  1.6.0  1.9.1.9000  local (/pecan/mod...
#>  PEcAn.data.hydrology   1.4.0  1.5.0  1.6.0  NA          NA                  
#>  PEcAn.data.land        1.4.0  1.5.0  1.6.0  1.9.0       local (/pecan/mod...
#>  PEcAn.data.mining      1.4.0  1.5.0  1.6.0  NA          NA                  
#>  PEcAn.data.remote      1.4.0  1.5.0  1.6.0  1.9.1       local (/pecan/mod...
#>  PEcAn.dvmdostem        <NA>   <NA>   1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.emulator         1.4.0  1.5.0  1.6.0  1.8.2       local (/pecan/mod...
#>  PEcAn.logger           <NA>   <NA>   1.6.0  1.8.4       local (/pecan/bas...
#>  PEcAn.photosynthesis   1.4.0  1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.priors           1.4.0  1.5.0  1.6.0  1.7.5       local (/pecan/mod...
#>  PEcAn.qaqc             1.4.0  1.5.0  1.6.0  1.7.5       local (/pecan/bas...
#>  PEcAn.remote           <NA>   <NA>   1.6.0  1.9.1       local (/pecan/bas...
#>  PEcAn.rtm              1.4.0  <NA>   <NA>   NA          NA                  
#>  PEcAn.settings         1.4.0  1.5.0  1.6.0  1.9.1.9000  local (/pecan/bas...
#>  PEcAn.uncertainty      1.4.0  1.5.0  1.6.0  1.9.0.9000  local (/pecan/mod...
#>  PEcAn.utils            1.4.0  1.5.0  1.6.0  1.8.2       local (/pecan/bas...
#>  PEcAn.visualization    1.4.0  1.5.0  1.6.0  1.8.2       local (/pecan/bas...
#>  PEcAn.workflow         <NA>   <NA>   1.6.0  1.10.0      local (/pecan/bas...
#>  PEcAnAssimSequential   <NA>   <NA>   <NA>   1.10.0.9000 local (/pecan/mod...
#>  PEcAnRTM               <NA>   1.5.0  1.6.0  1.9.1       local (/pecan/mod...

# multiple matches are possible when exact = FALSE
pecan_version("v1.3", exact = TRUE)
#>  package                v1.3  installed   source              
#>  PEcAn                  1.2.6 NA          NA                  
#>  PEcAn.BASGRA           <NA>  1.8.2       local (/pecan/mod...
#>  PEcAn.BIOCRO           1.2   1.7.5       local (/pecan/mod...
#>  PEcAn.CLM45            <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.DALEC            <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.DB               1.2.6 1.8.2.9000  local (/pecan/bas...
#>  PEcAn.ED               1.2.6 NA          NA                  
#>  PEcAn.ED2              <NA>  1.8.2       local (/pecan/mod...
#>  PEcAn.FATES            <NA>  1.8.1       local (/pecan/mod...
#>  PEcAn.GDAY             <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.JULES            <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.LDNDC            <NA>  1.0.2       local (/pecan/mod...
#>  PEcAn.LINKAGES         <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.LPJGUESS         <NA>  1.9.0       local (/pecan/mod...
#>  PEcAn.MA               1.2.6 1.7.5.9000  local (/pecan/mod...
#>  PEcAn.MAAT             <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.MAESPA           <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.ModelName        1.2   0.0.1       local (/pecan/mod...
#>  PEcAn.RothC            <NA>  0.0.0.9000  local (/pecan/mod...
#>  PEcAn.SIBCASA          <NA>  0.0.3       local (/pecan/mod...
#>  PEcAn.SIPNET           1.2.6 1.10.0      local (/pecan/mod...
#>  PEcAn.STICS            <NA>  1.8.2       local (/pecan/mod...
#>  PEcAn.all              1.2.6 1.10.0      local               
#>  PEcAn.allometry        <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.assim.batch      1.2.6 1.9.1       local (/pecan/mod...
#>  PEcAn.assim.sequential 1.2.6 NA          NA                  
#>  PEcAn.benchmark        <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.data.atmosphere  1.2.6 1.9.1.9000  local (/pecan/mod...
#>  PEcAn.data.land        1.2.6 1.9.0       local (/pecan/mod...
#>  PEcAn.data.remote      <NA>  1.9.1       local (/pecan/mod...
#>  PEcAn.dvmdostem        <NA>  1.7.5       local (/pecan/mod...
#>  PEcAn.emulator         1.0   1.8.2       local (/pecan/mod...
#>  PEcAn.logger           <NA>  1.8.4       local (/pecan/bas...
#>  PEcAn.photosynthesis   1.2.6 1.7.5       local (/pecan/mod...
#>  PEcAn.priors           1.2.6 1.7.5       local (/pecan/mod...
#>  PEcAn.qaqc             1.0   1.7.5       local (/pecan/bas...
#>  PEcAn.remote           <NA>  1.9.1       local (/pecan/bas...
#>  PEcAn.rtm              1.2.6 NA          NA                  
#>  PEcAn.settings         <NA>  1.9.1.9000  local (/pecan/bas...
#>  PEcAn.uncertainty      1.2.6 1.9.0.9000  local (/pecan/mod...
#>  PEcAn.utils            1.2.6 1.8.2       local (/pecan/bas...
#>  PEcAn.visualization    1.2.6 1.8.2       local (/pecan/bas...
#>  PEcAn.workflow         <NA>  1.10.0      local (/pecan/bas...
#>  PEcAnAssimSequential   <NA>  1.10.0.9000 local (/pecan/mod...
#>  PEcAnRTM               <NA>  1.9.1       local (/pecan/mod...
pecan_version("v1.3", exact = FALSE)
#>  package                v1.3  v1.3.2 v1.3.4 v1.3.5 v1.3.6 v1.3.6.1 installed  
#>  PEcAn                  1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    NA         
#>  PEcAn.BASGRA           <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.8.2      
#>  PEcAn.BIOCRO           1.2   1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.7.5      
#>  PEcAn.CLM45            <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.7.5      
#>  PEcAn.DALEC            <NA>  <NA>   <NA>   <NA>   <NA>   1.3.7    1.7.5      
#>  PEcAn.DB               1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.8.2.9000 
#>  PEcAn.ED               1.2.6 <NA>   <NA>   <NA>   <NA>   <NA>     NA         
#>  PEcAn.ED2              <NA>  1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.8.2      
#>  PEcAn.FATES            <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.8.1      
#>  PEcAn.GDAY             <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.7.5      
#>  PEcAn.JULES            <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.7.5      
#>  PEcAn.LDNDC            <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.0.2      
#>  PEcAn.LINKAGES         <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.7.5      
#>  PEcAn.LPJGUESS         <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.9.0      
#>  PEcAn.MA               1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.7.5.9000 
#>  PEcAn.MAAT             <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.7.5      
#>  PEcAn.MAESPA           <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.7.5      
#>  PEcAn.ModelName        1.2   1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    0.0.1      
#>  PEcAn.RothC            <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     0.0.0.9000 
#>  PEcAn.SIBCASA          <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     0.0.3      
#>  PEcAn.SIPNET           1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.10.0     
#>  PEcAn.STICS            <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.8.2      
#>  PEcAn.all              1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.10.0     
#>  PEcAn.allometry        <NA>  <NA>   1.3.3  1.3.3  1.3.3  1.3.7    1.7.5      
#>  PEcAn.assim.batch      1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.9.1      
#>  PEcAn.assim.sequential 1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    NA         
#>  PEcAn.benchmark        <NA>  <NA>   <NA>   1.3.3  1.3.3  1.3.7    1.7.5      
#>  PEcAn.dalec            <NA>  <NA>   1.3.1  1.3.1  1.3.1  <NA>     NA         
#>  PEcAn.data.atmosphere  1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.9.1.9000 
#>  PEcAn.data.hydrology   <NA>  <NA>   <NA>   <NA>   <NA>   1.3.7    NA         
#>  PEcAn.data.land        1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.9.0      
#>  PEcAn.data.mining      <NA>  <NA>   1.3.3  1.3.3  1.3.3  1.3.7    NA         
#>  PEcAn.data.remote      <NA>  <NA>   <NA>   <NA>   <NA>   1.3.7    1.9.1      
#>  PEcAn.dvmdostem        <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.7.5      
#>  PEcAn.emulator         1.0   1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.8.2      
#>  PEcAn.logger           <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.8.4      
#>  PEcAn.photosynthesis   1.2.6 1.3.1  1.3.3  1.3.3  1.3.6  1.3.7    1.7.5      
#>  PEcAn.priors           1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.7.5      
#>  PEcAn.qaqc             1.0   1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.7.5      
#>  PEcAn.remote           <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.9.1      
#>  PEcAn.rtm              1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    NA         
#>  PEcAn.settings         <NA>  1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.9.1.9000 
#>  PEcAn.uncertainty      1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.9.0.9000 
#>  PEcAn.utils            1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.8.2      
#>  PEcAn.visualization    1.2.6 1.3.1  1.3.3  1.3.3  1.3.3  1.3.7    1.8.2      
#>  PEcAn.workflow         <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.10.0     
#>  PEcAnAssimSequential   <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.10.0.9000
#>  PEcAnRTM               <NA>  <NA>   <NA>   <NA>   <NA>   <NA>     1.9.1      
#>  source              
#>  NA                  
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/bas...
#>  NA                  
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local               
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  NA                  
#>  local (/pecan/mod...
#>  NA                  
#>  local (/pecan/mod...
#>  NA                  
#>  local (/pecan/mod...
#>  NA                  
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/bas...
#>  local (/pecan/mod...
#>  local (/pecan/mod...
#>  local (/pecan/bas...
#>  local (/pecan/bas...
#>  NA                  
#>  local (/pecan/bas...
#>  local (/pecan/mod...
#>  local (/pecan/bas...
#>  local (/pecan/bas...
#>  local (/pecan/bas...
#>  local (/pecan/mod...
#>  local (/pecan/mod...